building-multimodal-ai-apps/requirements.txt
openai>=1.0.0
anthropic>=0.30.0
pydantic>=2.0.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
numpy>=1.24.0
